+++ data-transfer-object/src/main/java/org/grnet/cat/dtos/registry/actor/MotivationActorResponse.java
@@ -6,7 +6,6 @@ import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
 
 @Schema(name="MotivationActorResponse", description="This object represents a MotivationActorJunction item.")
-
 public class MotivationActorResponse {
 
     @Schema(