+++ data-transfer-object/src/main/java/org/grnet/cat/dtos/registry/template/TemplateResponse.java
@@ -1,10 +1,11 @@
-package org.grnet.cat.dtos.template;
+package org.grnet.cat.dtos.registry.template;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
 import org.grnet.cat.dtos.ActorDto;
 import org.grnet.cat.dtos.assessment.AssessmentTypeDto;
+import org.grnet.cat.dtos.template.TemplateDto;
 
 @Schema(name = "Template", description = "This object represents the Template.")
 public class TemplateResponse {